+++ apps/web/client/src/server/api/root.ts
@@ -1,5 +1,5 @@
 import { createCallerFactory, createTRPCRouter } from '~/server/api/trpc';
-import { canvasRouter, chatRouter, codeRouter, domainRouter, frameRouter, githubRouter, invitationRouter, memberRouter, projectRouter, sandboxRouter, settingsRouter, subscriptionRouter, usageRouter, userCanvasRouter, userRouter } from './routers';
+import { canvasRouter, chatRouter, codeRouter, domainRouter, frameRouter, githubRouter, invitationRouter, memberRouter, projectRouter, publishRouter, sandboxRouter, settingsRouter, subscriptionRouter, usageRouter, userCanvasRouter, userRouter } from './routers';
 
 /**
  * This is the primary router for your server.
@@ -22,6 +22,7 @@ export const appRouter = createTRPCRouter({
     github: githubRouter,
     subscription: subscriptionRouter,
     usage: usageRouter,
+    publish: publishRouter,
 });
 
 // export type definition of API
