+++ apps/web/client/src/server/api/routers/index.ts
@@ -4,6 +4,7 @@ export * from './forward';
 export * from './github';
 export * from './image';
 export * from './project';
+export * from './publish';
 export * from './subscription';
 export * from './usage';
 export * from './user';
